+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Add files/size and files/children
 * Add clj-toolbox.atoms
 * Add clj-toolbox.shelltools
+* Rename clj-toolbox.string-tools to clj-toolbox.strings
+* Add clj-toolbox.strings/split-whitespace
 
 ## [0.9.0] — 2024-12-30
 * Add property tests using clojure.test.check for streams and base64